requirements.txt
alipay-sdk-python==3.6.528
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
django-import-export==3.0.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
lxml==4.9.2
MarkupPy==1.14
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.3.0
pyasn1==0.4.8
pycryptodome==3.16.0
PyMySQL==1.0.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
rsa==4.9
sqlparse==0.4.3
tablib==3.3.0
tzdata==2022.6
urllib3==1.26.13
xlrd==2.0.1
xlwt==1.3.0
